updated forcaster to use the new met.no api. yr.no will be end of life in december. installer can now be invoked with update which does not reboot after install/update of dependencies. added latlong.json file for new weather api. core and airdata now use forcast2.0.py for weather data.

+++ install.py
@@ -20,13 +20,15 @@ dir = os.getcwd()
 fstab_comment = "#temp filesystem only in RAM for use on Airiana tempfiles.\n"
 fstab_cmd = "tmpfs "+dir+"/RAM tmpfs defaults,noatime,nosuid,mode=0755,size=50m 0 0\n"
 # MAKE RAM DRIve for linux logs var/logs
-fstab_cmd += "tmpfs /var/log tmpfs defaults,noatime,mode=0755,size=75m 0 0/n" 
+fstab_cmd += "tmpfs /home/pi/airiana/RAM tmpfs defaults,noatime,nosuid,mode=0755,uid=pi,gid=pi,size=50m 0 0"
+
 # INSTALL DEPS#
 os.system("apt-get -y install python-matplotlib")
 os.system("pip install  minimalmodbus")
 os.system("pip install  progressbar")
 os.system("pip install requests")
 os.system("pip install pyModbusTCP")
+os.system("pip install pyephem")
 os.system("apt-get -y install ntp")
 # NEED TO SET LOCALE
 os.system("cp ./systemfiles/timezone /etc/")
@@ -49,21 +51,23 @@ else:
 	 fstab_file.write(fstab_cmd)
 	 fstab_file.close()
 #install system services for airaina and controller
-if not os.path.lexist("/etc/systemd/system/airiana.service"):
+if not os.path.lexists("/etc/systemd/system/airiana.service") or "update" in sys.argv:
 	os.system("cp ./systemfiles/airiana.service /etc/systemd/system/")
-if not os.path.lexist("/etc/systemd/system/controller.service"):
+if not os.path.lexists("/etc/systemd/system/controller.service")) or "update" in sys.argv:
 	os.system("cp ./systemfiles/controller.service /etc/systemd/system/")
 
 os.system("systemctl enable airiana.service")
 os.system("systemctl enable controller.service")
 # redir console from uart
 boot_cmd= "dwc_otg.lpm_enable=0 console=tty1 root=/dev/mmcblk0p2 rootfstype=ext4 elevator=deadline fsck.repair=yes rootwait"
-os.system("echo " +boot_cmd+"> /boot/cmdline.txt" )
+if boot_cmd not in open("/boot/cmdline.txt").read():
+	print "adding boot cmdline config"
+	os.system("echo " +boot_cmd+"> /boot/cmdline.txt" )
 ##
 ## setup airiana as host#
 os.system ("cp ./systemfiles/hosts /etc/")
 #Touch a data log file
-os.system ("touch data.log.1")
+os.system ("touch data.log")
 
 os.chdir("./public/")
 # link files to RAM-disk
@@ -77,6 +81,7 @@ os.system("chown pi:pi ../RAM/*")
 tmp  =os.popen("crontab -u pi -l").read()
 os.system("echo \"" + tmp+ "0 */4 * * * sudo /usr/bin/python "+dir+"/updater.py\"|crontab -u pi -")
 
-print "Installation completed, reboot in 15 sec"
-os.system("sleep 15")
-os.system("reboot")
+if "update" not in sys.argv:
+	print "Installation completed, reboot in 15 sec"
+	os.system("sleep 15")
+	os.system("reboot")
